+++ app/src/main/kotlin/com/simplemobiletools/contacts/pro/dialogs/CreateNewGroupDialog.kt
@@ -10,9 +10,8 @@ import com.simplemobiletools.commons.extensions.toast
 import com.simplemobiletools.commons.extensions.value
 import com.simplemobiletools.commons.models.RadioItem
 import com.simplemobiletools.contacts.pro.R
-import com.simplemobiletools.contacts.pro.extensions.config
+import com.simplemobiletools.contacts.pro.extensions.getPrivateContactSource
 import com.simplemobiletools.contacts.pro.helpers.ContactsHelper
-import com.simplemobiletools.contacts.pro.helpers.SMT_PRIVATE
 import com.simplemobiletools.contacts.pro.models.ContactSource
 import com.simplemobiletools.contacts.pro.models.Group
 import kotlinx.android.synthetic.main.dialog_create_new_group.view.*
@@ -35,15 +34,9 @@ class CreateNewGroupDialog(val activity: BaseSimpleActivity, val callback: (newG
                             }
 
                             val contactSources = ArrayList<ContactSource>()
-                            if (activity.config.localAccountName.isNotEmpty()) {
-                                val localAccountName = activity.config.localAccountName
-                                contactSources.add(ContactSource(localAccountName, activity.config.localAccountType, localAccountName))
-                            }
-
                             ContactsHelper(activity).getContactSources {
                                 it.filter { it.type.contains("google", true) }.mapTo(contactSources) { ContactSource(it.name, it.type, it.name) }
-                                val phoneSecret = activity.getString(R.string.phone_storage_hidden)
-                                contactSources.add(ContactSource(phoneSecret, SMT_PRIVATE, phoneSecret))
+                                contactSources.add(activity.getPrivateContactSource())
 
                                 val items = ArrayList<RadioItem>()
                                 contactSources.forEachIndexed { index, contactSource ->