+++ app/src/main/kotlin/com/simplemobiletools/contacts/pro/extensions/Context.kt
@@ -218,7 +218,6 @@ fun Context.hasContactPermissions() = hasPermission(PERMISSION_READ_CONTACTS) &&
 
 fun Context.getPublicContactSource(source: String, callback: (String) -> Unit) {
     when (source) {
-        config.localAccountName -> callback(getString(R.string.phone_storage))
         SMT_PRIVATE -> callback(getString(R.string.phone_storage_hidden))
         else -> {
             Thread {
@@ -326,13 +325,16 @@ fun Context.getContactPublicUri(contact: Contact): Uri {
 }
 
 fun Context.getVisibleContactSources(): ArrayList<String> {
-    val sources = ContactsHelper(this).getDeviceContactSources()
-    val phoneSecret = getString(R.string.phone_storage_hidden)
-    sources.add(ContactSource(phoneSecret, SMT_PRIVATE, phoneSecret))
+    val sources = getAllContactSources()
     val ignoredContactSources = config.ignoredContactSources
-    val sourceNames = ArrayList(sources).filter { !ignoredContactSources.contains(it.getFullIdentifier()) }
-            .map { if (it.type == SMT_PRIVATE) SMT_PRIVATE else it.name }.toMutableList() as ArrayList<String>
-    return sourceNames
+    return ArrayList(sources).filter { !ignoredContactSources.contains(it.getFullIdentifier()) }
+            .map { it.name }.toMutableList() as ArrayList<String>
+}
+
+fun Context.getAllContactSources(): List<ContactSource> {
+    val sources = ContactsHelper(this).getDeviceContactSources()
+    sources.add(getPrivateContactSource())
+    return sources.toMutableList()
 }
 
 @TargetApi(Build.VERSION_CODES.N)
@@ -390,3 +392,5 @@ fun Context.deleteBlockedNumber(number: String) {
 
 @TargetApi(Build.VERSION_CODES.M)
 fun Context.isDefaultDialer() = isMarshmallowPlus() && telecomManager.defaultDialerPackage == packageName
+
+fun Context.getPrivateContactSource() = ContactSource(SMT_PRIVATE, SMT_PRIVATE, getString(R.string.phone_storage_hidden))