+++ app/src/main/kotlin/com/simplemobiletools/contacts/pro/helpers/ContactsHelper.kt
@@ -2,7 +2,6 @@ package com.simplemobiletools.contacts.pro.helpers
 
 import android.accounts.Account
 import android.accounts.AccountManager
-import android.annotation.SuppressLint
 import android.content.*
 import android.database.Cursor
 import android.graphics.Bitmap
@@ -10,7 +9,6 @@ import android.net.Uri
 import android.os.Environment
 import android.os.Handler
 import android.os.Looper
-import android.provider.CallLog
 import android.provider.ContactsContract
 import android.provider.ContactsContract.CommonDataKinds
 import android.provider.ContactsContract.CommonDataKinds.Nickname
@@ -34,11 +32,17 @@ class ContactsHelper(val context: Context) {
     private val BATCH_SIZE = 100
     private var displayContactSources = ArrayList<String>()
 
-    fun getContacts(callback: (ArrayList<Contact>) -> Unit) {
+    fun getContacts(ignoredContactSources: HashSet<String>? = null, callback: (ArrayList<Contact>) -> Unit) {
         Thread {
             val contacts = SparseArray<Contact>()
             displayContactSources = context.getVisibleContactSources()
-            getDeviceContacts(contacts)
+            if (ignoredContactSources != null) {
+                displayContactSources = context.getAllContactSources().filter {
+                    !ignoredContactSources.contains(it.getFullIdentifier())
+                }.map { it.getFullIdentifier() }.toMutableList() as ArrayList
+            }
+
+            getDeviceContacts(contacts, ignoredContactSources)
 
             if (displayContactSources.contains(SMT_PRIVATE)) {
                 LocalContactsHelper(context).getAllContacts().forEach {
@@ -52,7 +56,7 @@ class ContactsHelper(val context: Context) {
             val resultContacts = ArrayList<Contact>(contactsSize)
 
             (0 until contactsSize).filter {
-                if (showOnlyContactsWithNumbers) {
+                if (ignoredContactSources == null && showOnlyContactsWithNumbers) {
                     contacts.valueAt(it).phoneNumbers.isNotEmpty()
                 } else {
                     true
@@ -61,7 +65,7 @@ class ContactsHelper(val context: Context) {
                 contacts.valueAt(it)
             }
 
-            if (context.config.filterDuplicates) {
+            if (ignoredContactSources == null && context.config.filterDuplicates) {
                 tempContacts = tempContacts.distinctBy {
                     it.getHashToCompare()
                 } as ArrayList<Contact>
@@ -97,13 +101,20 @@ class ContactsHelper(val context: Context) {
     }
 
     private fun getContentResolverAccounts(): HashSet<ContactSource> {
-        val uri = ContactsContract.Data.CONTENT_URI
+        val sources = HashSet<ContactSource>()
+        arrayOf(ContactsContract.Groups.CONTENT_URI, ContactsContract.Settings.CONTENT_URI, ContactsContract.RawContacts.CONTENT_URI).forEach {
+            fillSourcesFromUri(it, sources)
+        }
+
+        return sources
+    }
+
+    private fun fillSourcesFromUri(uri: Uri, sources: HashSet<ContactSource>) {
         val projection = arrayOf(
                 ContactsContract.RawContacts.ACCOUNT_NAME,
                 ContactsContract.RawContacts.ACCOUNT_TYPE
         )
 
-        val sources = HashSet<ContactSource>()
         var cursor: Cursor? = null
         try {
             cursor = context.contentResolver.query(uri, projection, null, null, null)
@@ -124,16 +135,14 @@ class ContactsHelper(val context: Context) {
         } finally {
             cursor?.close()
         }
-
-        return sources
     }
 
-    private fun getDeviceContacts(contacts: SparseArray<Contact>) {
+    private fun getDeviceContacts(contacts: SparseArray<Contact>, ignoredContactSources: HashSet<String>?) {
         if (!context.hasContactPermissions()) {
             return
         }
 
-        val ignoredSources = context.config.ignoredContactSources
+        val ignoredSources = ignoredContactSources ?: context.config.ignoredContactSources
         val uri = ContactsContract.Data.CONTENT_URI
         val projection = getContactProjection()
 
@@ -676,7 +685,7 @@ class ContactsHelper(val context: Context) {
         return groups
     }
 
-    fun getDeviceStoredGroups(): ArrayList<Group> {
+    private fun getDeviceStoredGroups(): ArrayList<Group> {
         val groups = ArrayList<Group>()
         if (!context.hasContactPermissions()) {
             return groups
@@ -840,8 +849,7 @@ class ContactsHelper(val context: Context) {
 
     private fun getContactSourcesSync(): ArrayList<ContactSource> {
         val sources = getDeviceContactSources()
-        val phoneSecret = context.getString(R.string.phone_storage_hidden)
-        sources.add(ContactSource(phoneSecret, SMT_PRIVATE, phoneSecret))
+        sources.add(context.getPrivateContactSource())
         return ArrayList(sources)
     }
 
@@ -867,10 +875,6 @@ class ContactsHelper(val context: Context) {
             it.name.isNotEmpty() && it.type.isNotEmpty() && !accounts.contains(Account(it.name, it.type))
         }
         sources.addAll(contentResolverAccounts)
-
-        if (sources.isEmpty() && context.config.localAccountName.isEmpty() && context.config.localAccountType.isEmpty()) {
-            sources.add(ContactSource("", "", ""))
-        }
 
         return sources
     }
@@ -1249,7 +1253,7 @@ class ContactsHelper(val context: Context) {
             ContentProviderOperation.newInsert(ContactsContract.RawContacts.CONTENT_URI).apply {
                 withValue(ContactsContract.RawContacts.ACCOUNT_NAME, contact.source)
                 withValue(ContactsContract.RawContacts.ACCOUNT_TYPE, getContactSourceType(contact.source))
-                withValue(ContactsContract.RawContacts.DIRTY, false)
+                withValue(ContactsContract.RawContacts.DIRTY, 1)
                 operations.add(build())
             }
 
@@ -1555,14 +1559,6 @@ class ContactsHelper(val context: Context) {
             context.showErrorToast(e)
         }
     }
-
-    @SuppressLint("MissingPermission")
-    fun getRecents(callback: (ArrayList<RecentCall>) -> Unit) {
-        Thread {
-            val calls = ArrayList<RecentCall>()
-            if (!context.hasPermission(PERMISSION_WRITE_CALL_LOG) || !context.hasPermission(PERMISSION_READ_CALL_LOG)) {
-                callback(calls)
-                return@Thread
             }
 
             val blockedNumbers = context.getBlockedNumbers()