+++ CHANGELOG.md
@@ -1,6 +1,20 @@
 Changelog
 ==========
 
+Version 6.3.0 *(2019-01-24)*
+----------------------------
+
+ * Fixed local contact source not being properly shown
+ * Fixed some glitches related to exporting contacts
+
+Version 6.2.0 *(2019-01-06)*
+----------------------------
+
+ * Removed the Recents tab due to Googles' latest security policies
+ * Allow showing letters on the dialpad
+ * Fixed some contact filtering issues
+ * Couple other smaller improvements
+
 Version 6.1.2 *(2018-12-28)*
 ----------------------------